+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # (Unreleased)
 
 ## Enhancements
+
 * Adds `Query` and `Status` fields to `OrganizationMembershipListOptions` to allow filtering memberships by status or username by @sebasslash [#550](https://github.com/hashicorp/go-tfe/pull/550)
+*  Add `ListForWorkspace` method to `VariableSets` interface to enable fetching variable sets associated with a workspace. [#551](https://github.com/hashicorp/go-tfe/pull/551)
 
 # v1.10.0